requirements.txt
django~=4.2
psycopg2
djangorestframework~=3.14.0
drf-extensions~=0.7.1
result~=0.9.0
PyJWT~=2.6.0
requests~=2.28.2
python-dotenv~=1.0.0